+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
   - `frame-minimal-reversed-linking0` to `frame-minimal-backward0`
 
+  - `mix0` to `frame-minimal-hybrid0`
+
 ## 0.2.0
 
 - Base version